move Jetty WS API exceptions to separate package
