+++ jetty-websocket/websocket-jetty-api/src/main/java/module-info.java
@@ -20,6 +20,7 @@ module org.eclipse.jetty.websocket.jetty.api
 {
     exports org.eclipse.jetty.websocket.api;
     exports org.eclipse.jetty.websocket.api.annotations;
+    exports org.eclipse.jetty.websocket.api.exceptions;
     exports org.eclipse.jetty.websocket.api.extensions;
     exports org.eclipse.jetty.websocket.api.util;
