+++ jetty-websocket/websocket-jetty-api/src/main/java/org/eclipse/jetty/websocket/api/exceptions/MessageTooLargeException.java
@@ -16,33 +16,30 @@
 // ========================================================================
 //
 
-package org.eclipse.jetty.websocket.api;
+package org.eclipse.jetty.websocket.api.exceptions;
 
+import org.eclipse.jetty.websocket.api.StatusCode;
+
+/**
+ * Exception when a message is too large for the internal buffers occurs and should trigger a connection close.
+ *
+ * @see StatusCode#MESSAGE_TOO_LARGE
+ */
 @SuppressWarnings("serial")
-public class CloseException extends WebSocketException
+public class MessageTooLargeException extends CloseException
 {
-    private int statusCode;
-
-    public CloseException(int closeCode, String message)
+    public MessageTooLargeException(String message)
     {
-        super(message);
-        this.statusCode = closeCode;
+        super(StatusCode.MESSAGE_TOO_LARGE, message);
     }
 
-    public CloseException(int closeCode, String message, Throwable cause)
+    public MessageTooLargeException(String message, Throwable t)
     {
-        super(message, cause);
-        this.statusCode = closeCode;
+        super(StatusCode.MESSAGE_TOO_LARGE, message, t);
     }
 
-    public CloseException(int closeCode, Throwable cause)
+    public MessageTooLargeException(Throwable t)
     {
-        super(cause);
-        this.statusCode = closeCode;
-    }
-
-    public int getStatusCode()
-    {
-        return statusCode;
+        super(StatusCode.MESSAGE_TOO_LARGE, t);
     }
 }
