+++ jetty-websocket/websocket-jetty-api/src/main/java/org/eclipse/jetty/websocket/api/exceptions/UpgradeException.java
@@ -16,7 +16,7 @@
 // ========================================================================
 //
 
-package org.eclipse.jetty.websocket.api;
+package org.eclipse.jetty.websocket.api.exceptions;
 
 import java.net.URI;
