+++ jetty-websocket/websocket-jetty-api/src/main/java/org/eclipse/jetty/websocket/api/exceptions/WebSocketTimeoutException.java
@@ -16,30 +16,27 @@
 // ========================================================================
 //
 
-package org.eclipse.jetty.websocket.api;
-
-import org.eclipse.jetty.websocket.api.annotations.WebSocket;
+package org.eclipse.jetty.websocket.api.exceptions;
 
 /**
- * Indicating that the provided Class is not a valid WebSocket as defined by the API.
- * <p>
- * A valid WebSocket should do one of the following:
- * <ul>
- * <li>Implement {@link WebSocketListener}</li>
- * <li>Extend {@link WebSocketAdapter}</li>
- * <li>Declare the {@link WebSocket &#064;WebSocket} annotation on the type</li>
- * </ul>
+ * Exception thrown to indicate a connection I/O timeout.
  */
-@SuppressWarnings("serial")
-public class InvalidWebSocketException extends WebSocketException
+public class WebSocketTimeoutException extends WebSocketException
 {
-    public InvalidWebSocketException(String message)
+    private static final long serialVersionUID = -6145098200250676673L;
+
+    public WebSocketTimeoutException(String message)
     {
         super(message);
     }
 
-    public InvalidWebSocketException(String message, Throwable cause)
+    public WebSocketTimeoutException(String message, Throwable cause)
     {
         super(message, cause);
+    }
+
+    public WebSocketTimeoutException(Throwable cause)
+    {
+        super(cause);
     }
 }
