+++ jetty-websocket/websocket-jetty-common/src/main/java/org/eclipse/jetty/websocket/common/JettyWebSocketFrameHandler.java
@@ -507,27 +507,27 @@ public class JettyWebSocketFrameHandler implements FrameHandler
     public static Throwable convertCause(Throwable cause)
     {
         if (cause instanceof MessageTooLargeException)
-            return new org.eclipse.jetty.websocket.api.MessageTooLargeException(cause.getMessage(), cause);
+            return new org.eclipse.jetty.websocket.api.exceptions.MessageTooLargeException(cause.getMessage(), cause);
 
         if (cause instanceof ProtocolException)
-            return new org.eclipse.jetty.websocket.api.ProtocolException(cause.getMessage(), cause);
+            return new org.eclipse.jetty.websocket.api.exceptions.ProtocolException(cause.getMessage(), cause);
 
         if (cause instanceof BadPayloadException)
-            return new org.eclipse.jetty.websocket.api.BadPayloadException(cause.getMessage(), cause);
+            return new org.eclipse.jetty.websocket.api.exceptions.BadPayloadException(cause.getMessage(), cause);
 
         if (cause instanceof CloseException)
-            return new org.eclipse.jetty.websocket.api.CloseException(((CloseException)cause).getStatusCode(), cause.getMessage(), cause);
+            return new org.eclipse.jetty.websocket.api.exceptions.CloseException(((CloseException)cause).getStatusCode(), cause.getMessage(), cause);
 
         if (cause instanceof WebSocketTimeoutException)
-            return new org.eclipse.jetty.websocket.api.WebSocketTimeoutException(cause.getMessage(), cause);
+            return new org.eclipse.jetty.websocket.api.exceptions.WebSocketTimeoutException(cause.getMessage(), cause);
 
         if (cause instanceof InvalidSignatureException)
-            return new org.eclipse.jetty.websocket.api.InvalidWebSocketException(cause.getMessage(), cause);
+            return new org.eclipse.jetty.websocket.api.exceptions.InvalidWebSocketException(cause.getMessage(), cause);
 
         if (cause instanceof UpgradeException)
         {
             UpgradeException ue = (UpgradeException)cause;
-            return new org.eclipse.jetty.websocket.api.UpgradeException(ue.getRequestURI(), ue.getResponseStatusCode(), cause);
+            return new org.eclipse.jetty.websocket.api.exceptions.UpgradeException(ue.getRequestURI(), ue.getResponseStatusCode(), cause);
         }
 
         return cause;
